Create MonthlyWorker and schedule DesiringGodWebsite retrieval
Created MonthlyWorker and abstracted concept of a scheduled worker to an
elixir behaviour. Each worker (daily and monthly) need only define what
frequency they want to schedule. To do so, they must define a callback
function schedule/0 that returns a time in seconds to repeat the "work"
to be done by the worker.